Fix spigot compatibility with adventure component in PlayerAdvancementDoneEvent
